### 2023-11-15 #6015 @rmdocherty
Hi,
In terms of reviewer suggestions, user haesleinhuepf who reviewed the SIAL paper above does good work on segmentation and has used SAM before, though doesn’t appear on the list of people who’ve agreed to review.

For users who do appear on the list above, user pchalp has done deep learning and segmentation in Python and user pr4deepr has done work with tools like ImageJ, napari and cellpose which are similar to this tool.
