### 2023-12-18 #6159 @editorialbot
[REVIEW]: SAMBA: A Trainable Segmentation Web-App with Smart Labelling
**Submitting author:** @rmdocherty (<a href="http://orcid.org/0000-0002-7332-0924">Ronan Docherty</a>)
**Repository:** https://github.com/tldr-group/samba-web
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
