### 2024-01-18 #6159 @jingpengw
Sorry, for the late response. I have completed the check list and have some other comments.

- the manual and video tutorial link in the web interface is not working. The [manual markdown file](https://github.com/tldr-group/samba-web/blob/main/MANUAL.md) does exist though.
-  The reference is not complete. For example: https://github.com/haesleinhuepf/napari-accelerated-pixel-and-object-classification
 https://github.com/fractal-napari-plugins-collection/napari-feature-classifier
