### 2024-01-18 d7f3335
added references to feature classifiers in the napari ecosystem (and napari itsel)
