+++ paper/paper.md
@@ -33,7 +33,7 @@ Segmentation is the assigning of a semantic class to every pixel in an image and
 
 # Statement of need
 
-Trainable segmentation tools like ilastik [@ilastik] and Trainable Weka Segmentation [@weka] are popular options for image segmentation in biology and materials science due to their flexibility. Both apply common image processing operations like blurs, edge detection, and texture filters to create a feature stack, then train a classifier (usually a random forest ensemble) to map from the stack to user drawn labels on a pixelwise basis. ilastix offers fast segmentations and a responsive live-view, whereas Trainable Weka Segmentation is part of and interoperates with the FIJI ImageJ library [@FIJI].
+Trainable segmentation tools like ilastik [@ilastik] and Trainable Weka Segmentation [@weka] are popular options for image segmentation in biology and materials science due to their flexibility. They apply common image processing operations like blurs, edge detection, and texture filters to create a feature stack, then train a classifier (usually a random forest ensemble) to map from the stack to user drawn labels on a pixelwise basis. ilastix offers fast segmentations and a responsive live-view, whereas Trainable Weka Segmentation is part of and interoperates with the FIJI ImageJ library [@FIJI]. Other trainable segmenters are based on the napari [@napari] image viewer, including napari-feature-classifier [@napari-feature-classifier] and the GPU-accelerated napari-apoc [@napari-apoc].
 
 Accurate, representative labels are important for segmentation quality, but tooling is limited to polygons and/or pixel brushes which can make labelling boundaries time-consuming. `SAMBA` (Segment Anything Model Based App) uses the recent object-detecting foundation model Segment Anything Model [@SAM] as a 'Smart Labelling' tool, which offers suggestions for associated objects as the cursor is moved which can then be added as labels. This 'Smart Labelling' can quickly pick out portions of the same phase and is resilient against noise, varying exposure, and certain types of artefacts *etc.* SAM has been applied to cell microscopy annotation and segmentation [@micro-SAM], but not to multi-phase segmentation.
 
