+++ paper/paper.bib
@@ -27,6 +27,96 @@
 }
 
 
+@software{napari,
+  author       = {Ahlers, Jannis and
+                  Althviz Moré, Daniel and
+                  Amsalem, Oren and
+                  Anderson, Ashley and
+                  Bokota, Grzegorz and
+                  Boone, Peter and
+                  Bragantini, Jordão and
+                  Buckley, Genevieve and
+                  Burt, Alister and
+                  Bussonnier, Matthias and
+                  Can Solak, Ahmet and
+                  Caporal, Clément and
+                  Doncila Pop, Draga and
+                  Evans, Kira and
+                  Freeman, Jeremy and
+                  Gaifas, Lorenzo and
+                  Gohlke, Christoph and
+                  Gunalan, Kabilar and
+                  Har-Gil, Hagai and
+                  Harfouche, Mark and
+                  Harrington, Kyle I. S. and
+                  Hilsenstein, Volker and
+                  Hutchings, Katherine and
+                  Lambert, Talley and
+                  Lauer, Jessy and
+                  Lichtner, Gregor and
+                  Liu, Ziyang and
+                  Liu, Lucy and
+                  Lowe, Alan and
+                  Marconato, Luca and
+                  Martin, Sean and
+                  McGovern, Abigail and
+                  Migas, Lukasz and
+                  Miller, Nadalyn and
+                  Muñoz, Hector and
+                  Müller, Jan-Hendrik and
+                  Nauroth-Kreß, Christopher and
+                  Nunez-Iglesias, Juan and
+                  Pape, Constantin and
+                  Pevey, Kim and
+                  Peña-Castellanos, Gonzalo and
+                  Pierré, Andrea and
+                  Rodríguez-Guerra, Jaime and
+                  Ross, David and
+                  Royer, Loic and
+                  Russell, Craig T. and
+                  Selzer, Gabriel and
+                  Smith, Paul and
+                  Sobolewski, Peter and
+                  Sofiiuk, Konstantin and
+                  Sofroniew, Nicholas and
+                  Stansby, David and
+                  Sweet, Andrew and
+                  Vierdag, Wouter-Michiel and
+                  Wadhwa, Pam and
+                  Weber Mendonça, Melissa and
+                  Windhager, Jonas and
+                  Winston, Philip and
+                  Yamauchi, Kevin},
+  title        = {{napari: a multi-dimensional image viewer for 
+                   Python}},
+  month        = jul,
+  year         = {2023},
+  publisher    = {Zenodo},
+  version      = {v0.4.18},
+  doi          = {10.5281/zenodo.8115575},
+  url          = {https://doi.org/10.5281/zenodo.8115575}
+}
+
+
+@software{napari-apoc,
+  author = {Robert Haase and Lee, Dohyeon and Draga Doncila Pop and Laura Žigutytė},
+  title = {{haesleinhuepf/napari-accelerated-pixel-and-object-classification: 0.14.1}},
+  month = nov,
+  year = {2023},
+  publisher = {Zenodo},
+  version = {0.14.1},
+  doi = {10.5281/zenodo.10071078},
+  url = {https://doi.org/10.5281/zenodo.10071078}
+}
+
+
+@software{napari-feature-classifier,
+author = {Luethi, Joel and Hess, Maks},
+title = {{napari-feature-classifier: An interactive classifier plugin to use with label images and feature measurements}},
+url = {https://github.com/fractal-napari-plugins-collection/napari-feature-classifier}
+}
+
+
 @article{FIJI,
 	title = {Fiji: an open-source platform for biological-image analysis},
 	volume = {9},
