### 2024-01-19 #6159 @rmdocherty
@jingpengw, thanks for the comments! I've fixed the links to manual but have yet to fix the video link - I'll edit that once the video is uploaded and will let you know when that happens. I've also added the references you listed to the paper.
