### 2024-01-23 #6159 @pchlap
Congratulations on a great piece of work @rmdocherty. A really useful, well designed and well implemented tool.

@AJQuinn I have completed my review. All items in the checklist are satified in my opinion and I don't request any additional changes. This work certainly meets the criteria for inclusion in JOSS.
