### 2024-05-25 #6159 @Kevin-Mattheus-Moerman
@rmdocherty I am happy to proceed to process this work for acceptance in JOSS. Please check the above points for you for the post-review checklist. Are you able to work on those? Let me know if you have any questions.
In addition, please check the above reference check. If any of those potentially missing DOIs listed, have DOIs please add them. No problem if they truly do not have any. Thanks.
