### 2024-05-25 46d5df7
updated dois of two arxiv papers and added orcid ids for co-authors
