+++ paper/paper.bib
@@ -141,7 +141,8 @@ url = {https://github.com/fractal-napari-plugins-collection/napari-feature-class
       Journal = {ArXiv e-prints},
       url={https://arxiv.org/abs/2304.02643},
       archivePrefix={arXiv},
-      primaryClass={cs.CV}
+      primaryClass={cs.CV},
+      doi={10.48550/arXiv.2304.02643}
 }
 
 @article {micro-SAM,
@@ -161,7 +162,8 @@ url = {https://github.com/fractal-napari-plugins-collection/napari-feature-class
       Journal = {ArXiv e-prints},
       eprint={2302.05442},
       archivePrefix={arXiv},
-      primaryClass={cs.CV}
+      primaryClass={cs.CV},
+      doi={10.48550/arXiv.2302.05442}
 }
 
 
