### 2024-05-26 #6159 @rmdocherty
Hi @Kevin-Mattheus-Moerman, thanks for getting this orgranised!

I've updated the paper to add orcid affiliations for each author, and have added two DOIs to `paper.bib` for the two arXiv preprints, the other works don't have an associated DOI. I have cited the preprints as `@misc`, unsure how or if this will affect whether these new DOIs are detected - should I cite them as a `@article` instead?

The [release version](https://github.com/tldr-group/samba-web/releases/tag/v1.0.0) is v1.0.0

The Zenodo DOI is 10.5281/zenodo.11307100 and is [linked here](https://zenodo.org/records/11307100)

Let me know if there's anything else that needs doing
