### 2024-05-30 #6159 @Kevin-Mattheus-Moerman
@rmdocherty  below I have two minor points in relation to the paper:

- [ ] Please add `United Kingdom` to your affiliation
- [ ] Check if `generalizable` should be `generalisable` to be consistent with the rest of the text which uses British English
- [ ] Can you check if you need to fix any of the DOI issues raised? (it is okay if some references actually do not have a DOI)
