+++ paper/paper.bib
@@ -203,6 +203,8 @@ url = {https://github.com/fractal-napari-plugins-collection/napari-feature-class
          Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
  journal={Journal of Machine Learning Research},
  volume={12},
+ number={85},
  pages={2825--2830},
- year={2011}
+ year={2011},
+ url={http://jmlr.org/papers/v12/pedregosa11a.html}
 }
