### 2024-05-31 #6159 @rmdocherty
@Kevin-Mattheus-Moerman

- I've added `United Kingdom` to the affiliations
- I've changed `generalizable` to `generalisable`, and have updated `characterization` to `characterisation` (also in the summary) following the same logic
- I have checked the last three DOI issues, and they don't have an associated DOI. The first two are software tools with no DOI/zenodo/associated paper on their homepage, github or citation instructions and the last is a JMLR paper which doesn't come with a DOI. I have slightly updated the .bib entry of the last citation to include the paper url and number in order to be more consistent
